Introducing FirehoseQIOService

+++ src/main/java/com/fulmicotone/qio/components/accumulator/IQueueIOAccumulatorLengthFunction.java
@@ -3,5 +3,5 @@ package com.fulmicotone.qio.components.accumulator;
 import java.util.function.Function;
 
 @FunctionalInterface
-public interface IQueueIOAccumulatorLengthFunction<I> extends Function<I, Integer> {
+public interface IQueueIOAccumulatorLengthFunction<I> extends Function<I, Double> {
 }
